Create README.md
Further investigation needs to be applied to the use of emojis in the ways in which they may have been incorporated into the security circumvention for GitHub over the past few years. I think it’s also important to follow that lead to the creation of Mojo As a programming language most likely the suspects were looking for involved with both.